add: put relay reply with succesful hashes

+++ assign2/src/main/java/server/MembershipService.java
@@ -6,6 +6,7 @@ import communication.MulticastHandler;
 import message.JoinMessage;
 import message.Message;
 import message.PutRelayMessage;
+import message.PutRelayReply;
 import server.state.InitNodeState;
 import server.state.NodeState;
 import server.tasks.ElectionTask;
@@ -183,21 +184,33 @@ public class MembershipService implements MembershipRMI {
         String joiningNodeHash = StoreUtils.sha256(joiningNode.id().getBytes(StandardCharsets.UTF_8));
         String thisNodeHash = StoreUtils.sha256(this.getStorageService()
                 .getNode().id().getBytes(StandardCharsets.UTF_8));
+        PutRelayMessage putMessage = new PutRelayMessage();
+
         for (String hash : this.getStorageService().getHashes()) {
             if (hash.compareTo(joiningNodeHash) >= 0 && hash.compareTo(thisNodeHash) <= 0) {
                 continue;
             }
-            PutRelayMessage putMessage = new PutRelayMessage();
             try {
                 File file = new File(this.getStorageService().getValueFilePath(hash));
                 byte[] bytes = Files.readAllBytes(file.toPath());
-                String key = StoreUtils.sha256(bytes);
-                putMessage.addValue(key, bytes);
+                putMessage.addValue(hash, bytes);
             } catch (IOException e) {
                 throw new IllegalArgumentException("File not found");
             }
-            CommunicationUtils.dispatchMessageToNodeWithoutReply(joiningNode, putMessage);
-            this.getStorageService().delete(hash);
+        }
+
+        if(putMessage.getValues().size() == 0) {
+            return;
+        }
+
+        try {
+            PutRelayReply putReply = (PutRelayReply) CommunicationUtils.dispatchMessageToNode(joiningNode, putMessage);
+
+            for(String hash : putReply.getSuccessfulHashes()) {
+                this.getStorageService().delete(hash);
+            }
+        } catch (ClassCastException e) {
+            e.printStackTrace();
         }
     }
 
@@ -208,18 +221,29 @@ public class MembershipService implements MembershipRMI {
             return;
         }
 
+        PutRelayMessage putMessage = new PutRelayMessage();
         for (String hash : getStorageService().getHashes()) {
-            PutRelayMessage putMessage = new PutRelayMessage();
             try {
                 File file = new File(getStorageService().getValueFilePath(hash));
                 byte[] bytes = Files.readAllBytes(file.toPath());
-                String key = StoreUtils.sha256(bytes);
-                putMessage.addValue(key, bytes);
+                putMessage.addValue(hash, bytes);
             } catch (IOException e) {
                 throw new IllegalArgumentException("File not found");
             }
-            CommunicationUtils.dispatchMessageToNode(successorNode, putMessage, null);
-            this.getStorageService().delete(hash);
+        }
+
+        if(putMessage.getValues().size() == 0) {
+            return;
+        }
+
+        try {
+            PutRelayReply putReply = (PutRelayReply) CommunicationUtils.dispatchMessageToNode(successorNode, putMessage);
+
+            for(String hash : putReply.getSuccessfulHashes()) {
+                this.getStorageService().delete(hash);
+            }
+        } catch (ClassCastException e) {
+            e.printStackTrace();
         }
     }
 }
